foundry-test

+++ contracts/interfaces/WETH9_.sol
@@ -4,5 +4,5 @@ pragma solidity 0.8.17;
 interface WETH9_ {
     function deposit() external payable;
 
-    function withdraw(uint wad) external;
+    function withdraw(uint256 wad) external;
 }
